ref(kafka): Move functionality into separate crate.
Create `relay-kafka` crate with its own configuration and producer
functionality.

Having dedicated crate will help to keep all the related code together
and it also simplifies the dependency tree.
