+++ relay-config/Cargo.toml
@@ -19,6 +19,7 @@ human-size = "0.4.1"
 num_cpus = "1.13.0"
 relay-auth = { path = "../relay-auth" }
 relay-common = { path = "../relay-common" }
+relay-kafka = { path = "../relay-kafka" }
 relay-log = { path = "../relay-log", features = ["init"] }
 relay-metrics = { path = "../relay-metrics" }
 relay-redis = { path = "../relay-redis" }
