+++ relay-server/Cargo.toml
@@ -19,6 +19,7 @@ processing = [
     "rdkafka-sys",
     "rdkafka-sys/cmake-build",
     "relay-config/processing",
+    "relay-kafka/processing",
     "relay-quotas/redis",
     "relay-redis/impl",
     "symbolic-unreal",
@@ -55,6 +56,7 @@ relay-common = { path = "../relay-common" }
 relay-config = { path = "../relay-config" }
 relay-filter = { path = "../relay-filter" }
 relay-general = { path = "../relay-general" }
+relay-kafka = { path = "../relay-kafka", optional = true }
 relay-log = { path = "../relay-log" }
 relay-metrics = { path = "../relay-metrics" }
 relay-profiling = { path = "../relay-profiling" }
