+++ relay-server/src/actors/store.rs
@@ -6,15 +6,14 @@ use std::sync::Arc;
 use std::time::Instant;
 
 use bytes::Bytes;
-use failure::{Fail, ResultExt};
+use failure::Fail;
 use once_cell::sync::OnceCell;
-use rdkafka::{error::KafkaError, producer::BaseRecord, ClientConfig};
-use rmp_serde::encode::Error as RmpError;
 use serde::{ser::Error, Serialize};
 
 use relay_common::{ProjectId, UnixTimestamp, Uuid};
-use relay_config::{Config, KafkaConfig, KafkaParams, KafkaTopic};
+use relay_config::Config;
 use relay_general::protocol::{self, EventId, SessionAggregates, SessionStatus, SessionUpdate};
+use relay_kafka::{config::KafkaTopic, Message, Producer, ProducerError};
 use relay_log::LogError;
 use relay_metrics::{Bucket, BucketValue, MetricNamespace, MetricResourceIdentifier};
 use relay_quotas::Scoping;
@@ -23,8 +22,7 @@ use relay_system::{AsyncResponse, FromMessage, Interface, Sender, Service};
 
 use crate::envelope::{AttachmentType, Envelope, Item, ItemType};
 use crate::service::{ServerError, ServerErrorKind};
-use crate::statsd::{RelayCounters, RelayHistograms};
-use crate::utils::{CaptureErrorContext, ThreadedProducer};
+use crate::statsd::RelayCounters;
 
 /// The maximum number of individual session updates generated for each aggregate item.
 const MAX_EXPLODED_SESSIONS: usize = 100;
@@ -34,20 +32,11 @@ const UNNAMED_ATTACHMENT: &str = "Unnamed Attachment";
 
 #[derive(Fail, Debug)]
 pub enum StoreError {
-    #[fail(display = "failed to send kafka message")]
-    SendFailed(#[cause] KafkaError),
-    #[fail(display = "failed to serialize kafka message")]
-    InvalidMsgPack(#[cause] RmpError),
-    #[fail(display = "failed to serialize json message")]
-    InvalidJson(#[cause] serde_json::Error),
-    #[fail(display = "failed to find the shard range")]
-    InvalidKafkaShardRange,
+    #[fail(display = "failed to send the message to kafka")]
+    SendFailed(#[cause] ProducerError),
     #[fail(display = "failed to store event because event id was missing")]
     NoEventId,
 }
-
-/// Temporary map used to deduplicate kafka producers
-type ReusedProducersMap<'a> = BTreeMap<Option<&'a str>, Arc<ThreadedProducer>>;
 
 fn make_distinct_id(s: &str) -> Uuid {
     static NAMESPACE: OnceCell<Uuid> = OnceCell::new();
@@ -56,136 +45,6 @@ fn make_distinct_id(s: &str) -> Uuid {
 
     s.parse()
         .unwrap_or_else(|_| Uuid::new_v5(namespace, s.as_bytes()))
-}
-
-pub fn make_producer<'a>(
-    config: &'a Config,
-    reused_producers: &mut ReusedProducersMap<'a>,
-    kafka_topic: KafkaTopic,
-) -> Result<Producer, ServerError> {
-    let mut client_config = ClientConfig::new();
-    match config
-        .kafka_config(kafka_topic)
-        .context(ServerErrorKind::KafkaError)?
-    {
-        KafkaConfig::Single(KafkaParams {
-            topic_name,
-            config_name,
-            params,
-        }) => {
-            if let Some(producer) = reused_producers.get(&config_name) {
-                return Ok(Producer::Single {
-                    topic_name: topic_name.to_string(),
-                    producer: Arc::clone(producer),
-                });
-            }
-
-            for config_p in params {
-                client_config.set(config_p.name.as_str(), config_p.value.as_str());
-            }
-
-            let producer = Arc::new(
-                client_config
-                    .create_with_context(CaptureErrorContext)
-                    .context(ServerErrorKind::KafkaError)?,
-            );
-
-            reused_producers.insert(config_name, Arc::clone(&producer));
-            Ok(Producer::Single {
-                topic_name: topic_name.to_string(),
-                producer,
-            })
-        }
-        KafkaConfig::Sharded { shards, configs } => {
-            let mut producers = BTreeMap::new();
-            for (shard, kafka_params) in configs {
-                if let Some(producer) = reused_producers.get(&kafka_params.config_name) {
-                    let cached_producer = Arc::clone(producer);
-                    producers.insert(
-                        shard,
-                        (kafka_params.topic_name.to_string(), cached_producer),
-                    );
-                    continue;
-                }
-                for config_p in kafka_params.params {
-                    client_config.set(config_p.name.as_str(), config_p.value.as_str());
-                }
-                let producer = Arc::new(
-                    client_config
-                        .create_with_context(CaptureErrorContext)
-                        .context(ServerErrorKind::KafkaError)?,
-                );
-                reused_producers.insert(kafka_params.config_name, Arc::clone(&producer));
-                producers.insert(shard, (kafka_params.topic_name.to_string(), producer));
-            }
-            Ok(Producer::Sharded(ShardedProducer { shards, producers }))
-        }
-    }
-}
-
-/// This object containes the Kafka producer variants for single and sharded configurations.
-pub enum Producer {
-    Single {
-        topic_name: String,
-        producer: Arc<ThreadedProducer>,
-    },
-    Sharded(ShardedProducer),
-}
-
-/// Sharded producer configuration.
-pub struct ShardedProducer {
-    /// The maximum number of shards for this producer.
-    shards: u64,
-    /// The actual Kafka producer assigned to the range of logical shards, where the `u64` in the map is
-    /// the inclusive beginning of the range.
-    producers: BTreeMap<u64, (String, Arc<ThreadedProducer>)>,
-}
-
-impl ShardedProducer {
-    /// Returns topic name and the Kafka producer based on the provided sharding key.
-    /// Returns error [`StoreError::InvalidKafkaShardRange`] if the shard range for the provided sharding
-    /// key could not be found.
-    pub fn get_producer(&self, sharding_key: u64) -> Result<(&str, &ThreadedProducer), StoreError> {
-        let shard = sharding_key % self.shards;
-        let (topic_name, producer) = self
-            .producers
-            .iter()
-            .take_while(|(k, _)| *k <= &shard)
-            .last()
-            .map(|(_, v)| v)
-            .ok_or(StoreError::InvalidKafkaShardRange)?;
-
-        Ok((topic_name, producer))
-    }
-}
-
-impl Producer {
-    /// Sends the message to Kafka using correct producer for the current topic.
-    fn send(&self, organization_id: u64, message: KafkaMessage) -> Result<(), StoreError> {
-        let serialized = message.serialize()?;
-        metric!(
-            histogram(RelayHistograms::KafkaMessageSize) = serialized.len() as u64,
-            variant = message.variant()
-        );
-        let key = message.key();
-        let (topic_name, producer) = match self {
-            Self::Single {
-                topic_name,
-                producer,
-            } => (topic_name.as_str(), producer.as_ref()),
-
-            Self::Sharded(sharded) => sharded.get_producer(organization_id)?,
-        };
-        let record = BaseRecord::to(topic_name).key(&key).payload(&serialized);
-
-        producer.send(record).map_err(|(kafka_error, _message)| {
-            relay_log::with_scope(
-                |scope| scope.set_tag("variant", message.variant()),
-                || relay_log::error!("error sending kafka message: {}", kafka_error),
-            );
-            StoreError::SendFailed(kafka_error)
-        })
-    }
 }
 
 struct Producers {
@@ -224,35 +83,69 @@ impl Producers {
     pub fn create(config: &Arc<Config>) -> Result<Self, ServerError> {
         let mut reused_producers = BTreeMap::new();
         let producers = Producers {
-            attachments: make_producer(&**config, &mut reused_producers, KafkaTopic::Attachments)?,
-            events: make_producer(&**config, &mut reused_producers, KafkaTopic::Events)?,
-            transactions: make_producer(
-                &**config,
+            attachments: Producer::create(
+                &config
+                    .kafka_config(KafkaTopic::Attachments)
+                    .map_err(|_| ServerErrorKind::KafkaError)?,
                 &mut reused_producers,
-                KafkaTopic::Transactions,
-            )?,
-            sessions: make_producer(&**config, &mut reused_producers, KafkaTopic::Sessions)?,
-            metrics_sessions: make_producer(
-                &**config,
+            )
+            .map_err(|_| ServerErrorKind::KafkaError)?,
+            events: Producer::create(
+                &config
+                    .kafka_config(KafkaTopic::Events)
+                    .map_err(|_| ServerErrorKind::KafkaError)?,
                 &mut reused_producers,
-                KafkaTopic::MetricsSessions,
-            )?,
-            metrics_transactions: make_producer(
-                &**config,
+            )
+            .map_err(|_| ServerErrorKind::KafkaError)?,
+            transactions: Producer::create(
+                &config
+                    .kafka_config(KafkaTopic::Transactions)
+                    .map_err(|_| ServerErrorKind::KafkaError)?,
                 &mut reused_producers,
-                KafkaTopic::MetricsTransactions,
-            )?,
-            profiles: make_producer(&**config, &mut reused_producers, KafkaTopic::Profiles)?,
-            replay_recordings: make_producer(
-                &**config,
+            )
+            .map_err(|_| ServerErrorKind::KafkaError)?,
+            sessions: Producer::create(
+                &config
+                    .kafka_config(KafkaTopic::Sessions)
+                    .map_err(|_| ServerErrorKind::KafkaError)?,
                 &mut reused_producers,
-                KafkaTopic::ReplayRecordings,
-            )?,
-            replay_events: make_producer(
-                &**config,
+            )
+            .map_err(|_| ServerErrorKind::KafkaError)?,
+            metrics_sessions: Producer::create(
+                &config
+                    .kafka_config(KafkaTopic::MetricsSessions)
+                    .map_err(|_| ServerErrorKind::KafkaError)?,
                 &mut reused_producers,
-                KafkaTopic::ReplayEvents,
-            )?,
+            )
+            .map_err(|_| ServerErrorKind::KafkaError)?,
+            metrics_transactions: Producer::create(
+                &config
+                    .kafka_config(KafkaTopic::MetricsTransactions)
+                    .map_err(|_| ServerErrorKind::KafkaError)?,
+                &mut reused_producers,
+            )
+            .map_err(|_| ServerErrorKind::KafkaError)?,
+            profiles: Producer::create(
+                &config
+                    .kafka_config(KafkaTopic::Profiles)
+                    .map_err(|_| ServerErrorKind::KafkaError)?,
+                &mut reused_producers,
+            )
+            .map_err(|_| ServerErrorKind::KafkaError)?,
+            replay_recordings: Producer::create(
+                &config
+                    .kafka_config(KafkaTopic::ReplayRecordings)
+                    .map_err(|_| ServerErrorKind::KafkaError)?,
+                &mut reused_producers,
+            )
+            .map_err(|_| ServerErrorKind::KafkaError)?,
+            replay_events: Producer::create(
+                &config
+                    .kafka_config(KafkaTopic::ReplayEvents)
+                    .map_err(|_| ServerErrorKind::KafkaError)?,
+                &mut reused_producers,
+            )
+            .map_err(|_| ServerErrorKind::KafkaError)?,
         };
         Ok(producers)
     }
@@ -452,7 +345,9 @@ impl StoreService {
         message: KafkaMessage,
     ) -> Result<(), StoreError> {
         if let Some(producer) = self.producers.get(topic) {
-            producer.send(organization_id, message)?;
+            producer
+                .send_message(organization_id, &message)
+                .map_err(StoreError::SendFailed)?;
         }
 
         Ok(())
@@ -1103,7 +998,7 @@ enum KafkaMessage {
     ReplayRecordingChunk(ReplayRecordingChunkKafkaMessage),
 }
 
-impl KafkaMessage {
+impl Message for KafkaMessage {
     fn variant(&self) -> &'static str {
         match self {
             KafkaMessage::Event(_) => "event",
@@ -1142,18 +1037,18 @@ impl KafkaMessage {
     }
 
     /// Serializes the message into its binary format.
-    fn serialize(&self) -> Result<Vec<u8>, StoreError> {
+    fn serialize(&self) -> Result<Vec<u8>, ProducerError> {
         match self {
             KafkaMessage::Session(message) => {
-                serde_json::to_vec(message).map_err(StoreError::InvalidJson)
+                serde_json::to_vec(message).map_err(ProducerError::InvalidJson)
             }
             KafkaMessage::Metric(message) => {
-                serde_json::to_vec(message).map_err(StoreError::InvalidJson)
+                serde_json::to_vec(message).map_err(ProducerError::InvalidJson)
             }
             KafkaMessage::ReplayEvent(message) => {
-                serde_json::to_vec(message).map_err(StoreError::InvalidJson)
+                serde_json::to_vec(message).map_err(ProducerError::InvalidJson)
             }
-            _ => rmp_serde::to_vec_named(&self).map_err(StoreError::InvalidMsgPack),
+            _ => rmp_serde::to_vec_named(&self).map_err(ProducerError::InvalidMsgPack),
         }
     }
 }
